+++ src/main/java/Send.java
@@ -1,16 +1,16 @@
 
 public class Send {
 	Commission commission=new Commission();
-	public static int Amount;
-	public static int Receiver_Account;
+	public static long Amount;
+	public static long Receiver_Account;
 	
-	int get_Amount(int amount){
-		int result;
+	long get_Amount(long amount){
+		long result;
 		this.Amount=amount;
 		result=send_Amount();
 		return result;
 	}
-	int send_Amount(){
+	long send_Amount(){
 		commission.get_Commission(Amount);
 		this.Amount=commission.get_TotalAmount();
 		return this.Amount;
