+++ src/main/java/Withdraw.java
@@ -1,14 +1,14 @@
 
 public class Withdraw {
-	private static int Amount;
+	private static long Amount;
 	
-	int get_Amount(int amount){
-		int result;
+	long get_Amount(long amount){
+		long result;
 		this.Amount=amount;
 		result=send_Amount();
 		return result;
 	}
-	int send_Amount()
+	long send_Amount()
 	{
 		Commission commission=new Commission();
 		commission.get_Commission(Amount);
